fix: 실패 사유 리턴하도록 수정

+++ src/main/java/org/example/unpluguserservice/service/AuthService.java
@@ -29,6 +29,9 @@ public class AuthService {
 
     @Transactional
     public UserResponseDto createUser(UserRequestDto request, PasswordEncoder passwordEncoder){
+        if(userRepository.findByUsername(request.getUsername()).isPresent()){
+            return null;
+        }
         User user = User.builder()
                 .username(request.getUsername())
                 .name(request.getName())
@@ -41,25 +44,25 @@ public class AuthService {
     }
 
     @Transactional(readOnly = true)
-    public void checkUsableUsername(String username){
-        if(userRepository.findByUsername(username).isPresent()){
-            throw new RuntimeException("이미 존재하는 아이디입니다.");
-        }
+    public boolean checkUsableUsername(String username){
+        return userRepository.findByUsername(username).isEmpty();
     }
 
     @Transactional(readOnly = true)
-    public void checkUsableNickname(String nickname){
-        if (userRepository.findByNickname(nickname).isPresent()){
-            throw new RuntimeException("이미 존재하는 닉네임입니다.");
-        }
+    public boolean checkUsableNickname(String nickname){
+        return userRepository.findByNickname(nickname).isEmpty();
     }
 
     public LoginResponse login(String username, String password){
+        if(userRepository.findByUsername(username).isEmpty()){
+            return new LoginResponse("아이디가 올바르지 않습니다.", null, null);
+        }
+
         User user = userRepository.findByUsername(username)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
 
         if (!passwordEncoder.matches(password, user.getPassword())){
-            throw new IllegalArgumentException("비밀번호가 올바르지 않습니다.");
+            return new LoginResponse("비밀번호가 올바르지 않습니다.", null, null);
         }
 
         String accessToken = jwtUtil.generateAccessToken(user.getUsername());
@@ -68,28 +71,31 @@ public class AuthService {
         RefreshToken r = new RefreshToken(null, refreshToken, user.getUsername(), new Date(System.currentTimeMillis() + refreshExpMs));
         refreshTokenRepository.save(r);
 
-        return new LoginResponse(accessToken, refreshToken);
+        return new LoginResponse("로그인에 성공했습니다.", accessToken, refreshToken);
     }
 
     public LoginResponse refresh(String refreshToken){
+        if (refreshTokenRepository.findByToken(refreshToken).isEmpty()){
+            return new LoginResponse("Invalid Refresh Token ", null, refreshToken);
+        }
         RefreshToken token = refreshTokenRepository.findByToken(refreshToken)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid refresh token"));
 
         if (token.getExpiryDate().before(new Date())){
             refreshTokenRepository.delete(token);
-            throw new IllegalArgumentException("Refresh Token expired");
+            return new LoginResponse("Refresh Token expired", null, refreshToken);
         }
 
         User user = userRepository.findByUsername(token.getUsername())
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
 
         String newAccess = jwtUtil.generateAccessToken(user.getUsername());
-        return new LoginResponse(newAccess, refreshToken);
+        return new LoginResponse("success", newAccess, refreshToken);
     }
 
     public void logout(String refreshToken){
         refreshTokenRepository.findByToken(refreshToken).ifPresent(refreshTokenRepository::delete);
     }
 
-    public static record LoginResponse(String accessToken, String refreshToken){}
+    public static record LoginResponse(String data, String accessToken, String refreshToken){}
 }
